requirements.txt

# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0

# YouTube & Transcript Processing
youtube-transcript-api==1.2.1
yt-dlp==2024.12.3
huggingface_hub==0.23.0
# Vector Database & Embeddings
faiss-cpu==1.8.0
numpy==1.26.4

# LangChain Core & Community
langchain==0.2.6
langchain-community==0.2.6

# OpenAI Integration (Models & Embeddings)
openai==1.65.0
langchain-openai==0.1.8


# UI & Frontend
gradio==4.44.1

# Data Processing
pandas==2.1.4
pydantic==2.5.0

# Utilities
python-dateutil==2.8.2
urllib3==2.1.0
chardet==5.2.0

# Optional: For better logging
colorama==0.4.6
tqdm==4.66.1


